Added pImpl idiom to FmodAudioSystem
